[Gold V] Title: 등차수열과 쿼리, Time: 596 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/23888. 등차수열과 쿼리/등차수열과 쿼리.cc
@@ -17,7 +17,7 @@ int main(void) {
     int q; cin >> q;
 
     while(q-->0) {
-        int c, l, r; cin >> c >> l >> r;
+        ll c, l, r; cin >> c >> l >> r;
         if(c==1) {
             cout << a*(r-l+1)+((r-1)*r/2-(l-2)*(l-1)/2)*d << '\n';
         } else {
